refactor: globalfee antehandler
